+++ src/main/java/com/example/AD/mongoService/service/MongoGrupoService.java
@@ -32,7 +32,7 @@ public class MongoGrupoService {
      * @param grupoDTO el grupo dto
      */
     public void crearGrupo(GrupoDTO grupoDTO) {
-        Grupo grupo = new Grupo(grupoDTO.getNome(),
+        Grupo grupo = new Grupo(grupoDTO.getId(), grupoDTO.getNome(),
                 grupoDTO.getXenero(), grupoDTO.getDataFormacion());
         grupoRepository.save(grupo);
 
@@ -43,6 +43,10 @@ public class MongoGrupoService {
      * @return la lista de grupos
      */
     public List<Grupo> getListGrupo(){
+        List<Grupo> grupos = grupoRepository.findAll();
+        if(grupos.isEmpty()){
+            return null;
+        }
         return grupoRepository.findAll();
     }
 
@@ -52,6 +56,10 @@ public class MongoGrupoService {
      * @return el grupo o nada
      */
     public Grupo getListGrupoById(String id){
+        Grupo grupo = grupoRepository.findByid(id);
+        if(grupo == null){
+            return null;
+        }
         return grupoRepository.findByid(id);
     }
 
@@ -60,6 +68,10 @@ public class MongoGrupoService {
      * @param id el is del grupo
      */
     public void deleteByIdService(String id){
+        Grupo grupo = grupoRepository.findByid(id);
+        if(grupo == null){
+            throw new IdExcepcion("Este id no existe para borrar un grupo");
+        }
         grupoRepository.deleteById(id);
     }
 
@@ -69,7 +81,10 @@ public class MongoGrupoService {
      * @param grupoDTO el grupo con los datos actualizados
      */
     public void updateByIdService(String id, GrupoDTO grupoDTO){
-        Grupo grupo = grupoRepository.findById(id).orElseThrow(() -> new IdExcepcion("No existe el id del grupo"));
+        Grupo grupo = grupoRepository.findByid(id);
+        if(grupo == null){
+            throw new IdExcepcion("Este id no existe para actualizar un grupo");
+        }
         grupo.setNome(grupoDTO.getNome());
         grupo.setXenero(grupoDTO.getXenero());
         grupo.setDataFormacion(grupoDTO.getDataFormacion());
